requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# OpenAI & Environment
openai>=1.40.0
python-dotenv>=1.0.0

# Templating & Static Files
jinja2>=3.1.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP & Security
httpx>=0.25.0
python-multipart>=0.0.6

# Logging & Monitoring
structlog>=23.2.0

# Data Processing
pandas>=1.5.0

# Line Bot
line-bot-sdk>=3.0.0 